adedd "pages"
pages are static page that lives in pages/.  They are pretty
self-explanatory I believe.

They are managed like the posts, see pages.edn.

This also adds a "project" page (also liked on top of every page) and
some other pages linked from them.

The only thing is that I needed an hack for the HTTP part: replace
link to ".gmi" files to ".html".  For the moment is OK, but in the
future can be a pain.  I need to find a way to generate correctly link
from both the gemini part and the http part from the same gemtext file.